healthcare-app-v3/services/aof-service/requirements.txt
flask==2.3.2
psutil==5.9.6
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
nats-py==2.1.0
aioredis>=2.0.0